requirements.txt
alembic==1.13.1
alembic-autogen-check==1.1.1
bcrypt==4.1.2
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
elastic-transport==8.11.0
elasticsearch==8.11.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
